break out checkVariables
